+++ CupTaster/Resources/Assets.xcassets/onboarding-background.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "IMG_0357.jpeg",
+      "filename" : "IMG_0488.jpeg",
       "idiom" : "universal"
     }
   ],
