+++ CupTaster/Resources/ru.lproj/Localizable.strings
@@ -56,7 +56,6 @@
 "Sample name" = "Название образца";
 "Cup %d" = "Чашка %d";
 "Notes" = "Заметки";
-"You can add general information templates in settings" = "Вы можете добавить шаблоны общей информации в настройках";
 "Tap for more" = "Показать больше";
 
 // Settings
