+++ CupTaster/Sample/Evaluation/Header Value/Radio Value.swift
@@ -26,11 +26,12 @@ extension RadioEvaluationValueView {
         
         var body: some View {
             let cupsCount: Int = Int(intensityQC.group.sample.cupping.cupsCount)
+            let values: [CGFloat] = getValues(cupsCount: cupsCount)
             let selectedCupsCount: Int = Int(cupsCountQC.value).digits.reduce(0, +)
             let calculatedValue: Double = intensityQC.value * Double(selectedCupsCount) / Double(cupsCount) * 5 * -4
             
             HStack {
-                ForEach(getValues(cupsCount: cupsCount), id: \.self) { value in
+                ForEach(values, id: \.self) { value in
                     if value == calculatedValue {
                         Text(formatValue(value: value))
                             .bold()
