+++ CupTaster/Sample/Sample.swift
@@ -7,8 +7,12 @@
 
 import SwiftUI
 
+enum SampleAppearance {
+    case criteria, info
+}
+
 struct SampleView: View {
-    @Environment(\.managedObjectContext) private var moc
+    @Environment(\.managedObjectContext) var moc
     @ObservedObject var cuppingModel: CuppingModel
     @State var sample: Sample
     
@@ -31,225 +35,5 @@ struct SampleView: View {
         case .criteria: criteriaAppearance
         case .info: infoAppearance
         }
-    }
-}
-
-// MARK: Change appearance
-
-enum SampleAppearance {
-    case criteria, info
-}
-
-// MARK: All appearances
-
-extension SampleView {
-    private var criteriaAppearance: some View {
-        ScrollView(showsIndicators: false) {
-            LazyVStack(spacing: 0) {
-                ForEach(
-                    sample.qualityCriteriaGroups
-                        .sorted(by: { $0.configuration.ordinalNumber < $1.configuration.ordinalNumber })
-                ) { qcGroup in
-                    EvaluationGroupView(cuppingModel: cuppingModel, qcGroup: qcGroup)
-                }
-            }
-            .padding(.horizontal)
-            .padding(.bottom, 100) // toolbar
-            .resignKeyboardOnDragGesture() { try? moc.save() }
-        }
-        .clipped()
-    }
-}
-
-extension SampleView {
-    private var infoAppearance: some View {
-        Form {
-            RadarChart(sample: sample)
-            
-            GeneralInfoSectionView(sample: sample)
-            
-            Section {
-                Button("Delete", role: .destructive) {
-                    withAnimation {
-                        moc.delete(sample)
-                        
-                        let cupping: Cupping = sample.cupping
-                        let sortedSamples: [Sample] = cupping.getSortedSamples()
-                        if sortedSamples.count > 1 {
-                            if cuppingModel.selectedSampleIndex != 0 {
-                                self.cuppingModel.selectedSampleIndex! -= 1
-                                self.cuppingModel.selectedSample = sortedSamples[cuppingModel.selectedSampleIndex!]
-                            } else {
-                                self.cuppingModel.selectedSample = sortedSamples[cuppingModel.selectedSampleIndex! + 1]
-                            }
-                        } else {
-                            self.cuppingModel.selectedSampleIndex = 0
-                            self.cuppingModel.selectedSample = nil
-                            self.cuppingModel.sampleViewVisible = false
-                        }
-                        
-                        try? moc.save()
-                    }
-                }
-            } footer: {
-                Color.clear.frame(height: 100) // toolbar
-            }
-        }
-    }
-    
-    private struct GeneralInfoSectionView: View {
-        @Environment(\.managedObjectContext) private var moc
-        @FetchRequest(
-            entity: SampleGeneralInfo.entity(),
-            sortDescriptors: [NSSortDescriptor(keyPath: \SampleGeneralInfo.ordinalNumber, ascending: false)]
-        ) var sgiFields: FetchedResults<SampleGeneralInfo>
-        let sample: Sample
-        
-        @State private var newSGIFieldTitle: String = ""
-        @State private var newSGIFieldVisible: Bool = false
-        @FocusState private var newSGIFieldFocused: Bool
-        
-        var body: some View {
-            let addedSGIFields: [SampleGeneralInfo] = Array(sample.generalInfo)
-            let suggestedSGIFields: [FetchedResults<SampleGeneralInfo>.Element] = sgiFields.filter {
-                $0.sample == nil && !addedSGIFields.map { $0.title }.contains($0.title)
-            }
-            
-            ForEach(addedSGIFields) { sgiField in
-                Section(sgiField.title) {
-                    SGIFieldView(sampleGeneralInfo: sgiField)
-                }
-            }
-            .onDelete { offsets in
-                for index in offsets {
-                    moc.delete(addedSGIFields[index])
-                }
-                try? moc.save()
-            }
-            
-            Section {
-                if newSGIFieldVisible {
-                    HStack {
-                        Button {
-                            newSGIFieldTitle = ""
-                            withAnimation { newSGIFieldVisible = false }
-                        } label: {
-                            Image(systemName: "xmark.circle.fill")
-                                .contentShape(Rectangle())
-                        }
-                        
-                        TextField("Additional field", text: $newSGIFieldTitle) { addNewGIField() }
-                            .submitLabel(.done)
-                            .focused($newSGIFieldFocused, equals: true)
-                        
-                        Button { addNewGIField() } label: {
-                            Image(systemName: "checkmark.circle.fill")
-                                .contentShape(Rectangle())
-                        }
-                    }
-                    .buttonStyle(BorderlessButtonStyle())
-                    .onAppear { newSGIFieldFocused = true }
-                } else {
-                    Button {
-                        withAnimation { newSGIFieldVisible = true }
-                    } label: {
-                        Label("Additional field", systemImage: "plus")
-                            .submitLabel(.done)
-                    }
-                }
-                
-                ForEach(suggestedSGIFields) { sgiField in
-                    Button {
-                        let newSGIField: SampleGeneralInfo = SampleGeneralInfo(context: moc)
-                        newSGIField.title = sgiField.title
-                        newSGIField.ordinalNumber = Int16(sgiFields.filter({ $0.sample == sample }).count)
-                        newSGIField.sample = sample
-                        
-                        newSGIFieldTitle = ""
-                        newSGIFieldFocused = false
-                        newSGIFieldVisible = false
-                    } label: {
-                        Label {
-                            Text(sgiField.title)
-                                .foregroundColor(.gray)
-                        } icon: {
-                            Image(systemName: "plus")
-                                .foregroundColor(.accentColor)
-                        }
-                    }
-                }
-            } footer: {
-                if sgiFields.filter({ $0.sample == nil }).count == 0 {
-                    Text("You can add general information templates in settings")
-                }
-            }
-        }
-        
-        func addNewGIField() {
-            let newSGIField: SampleGeneralInfo = SampleGeneralInfo(context: moc)
-            newSGIField.title = newSGIFieldTitle
-            newSGIField.ordinalNumber = Int16(sgiFields.filter({ $0.sample == sample }).count)
-            newSGIField.sample = sample
-            
-            newSGIFieldTitle = ""
-            newSGIFieldFocused = false
-            newSGIFieldVisible = false
-        }
-    }
-    
-    private struct SGIFieldView: View {
-        @Environment(\.managedObjectContext) private var moc
-        @ObservedObject var sampleGeneralInfo: SampleGeneralInfo
-        
-        var body: some View {
-            TextField("", text: $sampleGeneralInfo.value) { try? moc.save() }
-                .submitLabel(.done)
-        }
-    }
-}
-
-// MARK: Extra
-
-extension SampleView {
-    public var preview: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            RadarChart(sample: sample, useShortLabels: true)
-            
-            HStack {
-                Group {
-                    if sample.finalScore != 0 { Text(String(format: "%.1f", sample.finalScore)) }
-                    else { Text("-") }
-                }
-                .lineLimit(1)
-                .frame(width: 30)
-                
-                Divider()
-                    .frame(height: 15)
-                
-                Text(sample.name)
-                    .fixedSize(horizontal: true, vertical: true)
-                
-                if sample.isFavorite {
-                    Spacer()
-                    Image(systemName: "heart.fill")
-                        .foregroundColor(.red)
-                }
-            }
-            .font(.caption)
-        }
-        .padding(10)
-        .background(Color(uiColor: .secondarySystemGroupedBackground))
-        .cornerRadius(15)
-    }
-}
-
-struct SampleFormRowView: View {
-    @Environment(\.managedObjectContext) private var moc
-    @ObservedObject var sample: Sample
-    
-    var body: some View {
-        TextField("Sample name", text: $sample.name)
-            .submitLabel(.done)
-            .onSubmit { try? moc.save() }
     }
 }
